test(core): move acceptance tests to same folder
