+++ packages/core/test/core/acceptance/loop.test.js
@@ -1,9 +1,9 @@
 'use strict';
 
 const { test, beforeEach, afterEach } = require('tap');
-const runner = require('../..').runner.runner;
-const { SSMS } = require('../../lib/ssms');
-const createTestServer = require('./targets/simple');
+const runner = require('../../..').runner.runner;
+const { SSMS } = require('../../../lib/ssms');
+const createTestServer = require('../targets/simple');
 
 let server;
 let port;
@@ -17,7 +17,7 @@ afterEach(() => {
 });
 
 test('simple loop', (t) => {
-  const script = require('./scripts/loop.json');
+  const script = require('../scripts/loop.json');
   script.config.target = `http://127.0.0.1:${port}`;
 
   runner(script).then(function (ee) {
@@ -42,7 +42,7 @@ test('simple loop', (t) => {
 });
 
 test('loop with range', (t) => {
-  const script = require('./scripts/loop_range.json');
+  const script = require('../scripts/loop_range.json');
   script.config.target = `http://127.0.0.1:${port}`;
 
   runner(script).then(function (ee) {
@@ -73,7 +73,7 @@ test('loop with range', (t) => {
 });
 
 test('loop with nested range', (t) => {
-  const script = require('./scripts/loop_nested_range.json');
+  const script = require('../scripts/loop_nested_range.json');
   script.config.target = `http://127.0.0.1:${port}`;
 
   runner(script).then(function (ee) {
